+++ packages/web-components/src/components/rux-log/test/rux-log.spec.tsx
@@ -45,7 +45,7 @@ describe('rux-log', () => {
                          <rux-table-header-cell class="rux-log__header-event-cell">
                            <div class="header-event-container">
                              Event
-                             <rux-input class="rux-log__filter" small="" type="search"></rux-input>
+                             <rux-input class="rux-log__filter" size="small" type="search"></rux-input>
                            </div>
                          </rux-table-header-cell>
                        </rux-table-header-row>