+++ packages/web-components/CHANGELOG.md
@@ -7,6 +7,12 @@
 -   Inputs with `type="password"` with now support toggle show/hide by including the `visibility` and `visibility-off` rux-icons as clickable.
 -   Password inputs now swap types between `password` and `text` to show/hide the password.
 
+*   `small` property has been removed. Use `size=small` instead.
+
+### Textarea
+
+-   `small` property has been removed. Use `rows` property instead to control the height of the element.
+
 ### Event names
 
 -   Gloably updated event names from camel case to lower case to be compatible with angular event name formats